+++ packages/components/src/Stack/Stack.tsx
@@ -12,8 +12,23 @@ import {
 // ---------------
 export interface StackProps extends AlignmentProp, GapSpaceProp {
   children?: ReactNode;
+
+  /**
+   * Stretch to fill space (vertical AND horizontal, useful if you want to change y alignment).
+   * @default false
+   */
   stretch?: boolean;
+
+  /**
+   * Horizontal alignment for the children.
+   * @default none
+   */
   alignX?: keyof typeof alignment.vertical.alignmentX;
+
+  /**
+   * Vertical alignment for the children.
+   * @default none
+   */
   alignY?: keyof typeof alignment.vertical.alignmentY;
 }
 
@@ -23,7 +38,6 @@ export const Stack = ({
   children,
   space = 0,
   stretch = false,
-  orientation,
   alignX,
   alignY,
   ...props